# contracts/openapi/plan-gateway.v1.yaml
openapi: 3.0.3
...
info:
  title: Overpass Plan Gateway
  version: 1.0.0
  description: |
    Read models over the collection plan.

    ## What this service is for

    `plan-gateway` is the eventually-consistent side of ADR-0003. It owns no
    business rules and makes no decisions: it projects events from `tasking`,
    `feasibility` and `planning` into read models shaped for the frontend, and
    serves them.

    That separation is the point. The planner's write path holds an invariant
    under a lock; this one answers "what does the globe look like" a few hundred
    milliseconds later, and the two must not contend.

    ## Staleness is exposed, never hidden

    Every projection response carries `as_of`, the stream position it was built
    from. A read model that cannot say how far behind it is forces the UI to
    guess, and the UI will guess optimistically — showing a plan as current when
    it is thirty seconds old and already superseded.

    ## Plans are addressed by bucket, not by time range

    This is the one genuinely contested decision in this document and it is
    recorded rather than assumed.

    A `CollectionPlan` exists per `(satellite_id, bucket_start)`, and
    `plan_version` increments within that pair. ADR-0012 made the projection
    append-only, so several versions of the same bucket coexist and a read must
    be able to name one.

    A free time range spanning two buckets would therefore match several plan
...

...

...

  contact:
    name: Overpass
...
servers:
...
paths:
  /v1/plans:
    get:
      operationId: listPlans
      tags: [plans]
      summary: List committed plans
      description: |
        Cursor-paginated, newest bucket first. Only the CURRENT version of each
        bucket is returned unless `include_superseded` is set — a client asking
        "what is planned" almost never means "and everything it replaced".
      parameters:
        - $ref: '#/components/parameters/SatelliteId'
        - name: bucket_start_after
          in: query
          description: Buckets starting at or after this instant.
          schema: { type: string, format: date-time }
        - name: bucket_start_before
          in: query
          schema: { type: string, format: date-time }
        - name: include_superseded
          in: query
          description: |
            Include plan versions that have been replaced. Off by default: the
            history is retained (ADR-0012) but it is not what a client asking
            for "the plan" wants, and returning it silently would double every
            response during active re-planning.
          schema: { type: boolean, default: false }
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of plans.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/plans/{satellite_id}/{bucket_start}:
    parameters:
      - $ref: '#/components/parameters/SatelliteIdPath'
      - name: bucket_start
        in: path
        required: true
        description: Start of the horizon bucket, RFC 3339.
        schema: { type: string, format: date-time }
    get:
      operationId: getPlan
      tags: [plans]
      summary: Get one plan
      description: |
        Returns the current version of this bucket's plan, or a named version.

        Addressed by bucket because `plan_version` only means something within
        one — see the note in this document's description.
      parameters:
        - name: plan_version
          in: query
          description: |
            A specific version. Omitted means the current one. Naming a version
            is how the retained history from ADR-0012 is actually reachable; a
            history nothing can address was not worth keeping.
          schema: { type: integer, minimum: 1 }
      responses:
        '200':
          description: The plan.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plan'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/acquisitions:
    get:
      operationId: listAcquisitions
      tags: [acquisitions]
      summary: List scheduled acquisitions
      description: |
        By time range, because an acquisition belongs to exactly one plan and
        needs no version arithmetic to interpret. This is the read behind the
        per-satellite timeline in M4-02.
      parameters:
        - $ref: '#/components/parameters/SatelliteId'
        - name: window_start
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: window_end
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: request_id
          in: query
          schema: { $ref: '#/components/schemas/Uuid' }
        - name: status
          in: query
          description: |
            Repeatable. Defaults to the live set — ACTIVE and EXECUTED —
            because a timeline showing superseded acquisitions alongside live
            ones is a timeline nobody can read.
          schema:
            type: array
            items:
              $ref: '#/components/schemas/AcquisitionStatus'
          style: form
          explode: true
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of acquisitions, earliest first.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcquisitionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/requests/{request_id}:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: getRequestView
      tags: [requests]
      summary: Request status, as the read model sees it
      description: |
        The projection, not the write model. `tasking-api` owns the request's
        authoritative state; this answers the different question of what the
        rest of the system currently believes, and says how stale that belief
        is.

        When a request lost a round, `unfulfilment` explains which constraint
        bound. That is the data behind the "why was my request rejected?" panel
        in M4-04, and it is structured rather than a message string because a
        string cannot be aggregated, charted, or acted on.
      responses:
        '200':
          description: The request view.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequestView'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/requests/{request_id}/opportunities:
    parameters:
      - $ref: '#/components/parameters/RequestId'
    get:
      operationId: listRequestOpportunities
      tags: [opportunities]
      summary: Candidates this request could have won
      description: |
        Every opportunity feasibility found, whether or not it was allocated.
        The losing candidates are what M4-05 renders as ghosts on the timeline —
        showing only the winner explains nothing about why it won.
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: A page of opportunities.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OpportunityPage'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/geo/plans/{satellite_id}/{bucket_start}/czml:
    parameters:
      - $ref: '#/components/parameters/SatelliteIdPath'
      - name: bucket_start
        in: path
        required: true
        schema: { type: string, format: date-time }
    get:
      operationId: getPlanCZML
      tags: [geo]
      summary: CZML rendering of one plan
      description: |
        A CZML document Cesium can load directly: the satellite's path over the
        bucket, and a footprint polygon per acquisition.

        **CZML is not specified here beyond its media type, deliberately.** It
        is a Cesium document format with its own structure outside JSON Schema,
        and a partial schema would be a contract that looks binding and is not.
        The shape we actually emit is pinned by golden-file tests in M1-15,
        which is a stronger guarantee than an incomplete declaration.
      parameters:
        - name: plan_version
          in: query
          schema: { type: integer, minimum: 1 }
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: A CZML document.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: array
                description: |
                  A CZML packet stream. The first packet is the document packet;
                  see https://github.com/AnalyticalGraphicsInc/czml-writer/wiki/CZML-Structure
                items:
                  type: object
                  additionalProperties: true
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/geo/satellites/czml:
    get:
      operationId: getConstellationCZML
      tags: [geo]
      summary: CZML rendering of the constellation's orbit tracks
      description: |
        A CZML document carrying a `position` and a `path` per satellite over a
        window, from the sampled ephemeris projected out of
        `feasibility.ephemeris.computed.v1`.

        **Independent of any plan, deliberately.** The per-plan CZML at
        `/v1/geo/plans/{satellite_id}/{bucket_start}/czml` renders the orbit of
        the one satellite that plan belongs to, which is the right answer when a
        plan is what you are looking at. It is the wrong answer for a globe:
        the constellation exists before the first plan is committed and after
        the last one is superseded, and a viewer who sees satellites appear only
        once something has been scheduled learns something false about the
        system.

        Empty is a legitimate answer, not an error. The ephemeris sweep runs on
        its own timer, so a window it has not reached yet has no samples — and a
        document with a clock and no satellites is the honest rendering of that.

        **CZML is not specified here beyond its media type**, for the same
        reason the per-plan document is not: it is a Cesium format with its own
        structure outside JSON Schema, and a partial schema would be a contract
        that looks binding and is not. The shape is pinned by golden files.
      parameters:
        - $ref: '#/components/parameters/SatelliteId'
        - name: window_start
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: window_end
          in: query
          required: true
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: A CZML packet stream, one satellite entity per satellite.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                type: array
                description: |
                  A CZML packet stream. The first packet is the document packet;
                  the rest are satellites.
                items:
                  type: object
                  additionalProperties: true
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /v1/geo/footprints:
    get:
      operationId: getFootprints
      tags: [geo]
      summary: GeoJSON footprints over a window
      description: |
        A FeatureCollection of acquisition footprints, for the deck.gl coverage
        view in M4-01.

        Geometry follows RFC 7946 and is not re-specified here. Coordinates are
        WGS84, longitude first — the same ordering as everywhere else in this
        system, and the one a swap silently relocates to another hemisphere.
      parameters:
        - $ref: '#/components/parameters/SatelliteId'
        - name: window_start
          in: query
          required: true
          schema: { type: string, format: date-time }
        - name: window_end
          in: query
          required: true
          schema: { type: string, format: date-time }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: |
            A GeoJSON FeatureCollection.

            Bounded by `limit`, and the collection reports whether it was
            truncated. An unbounded geometry endpoint is a denial-of-service
            waiting for a client to ask for a year; silently truncating instead
            would let a viewport show a coverage gap that does not exist, which
            is worse than an error.
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/geo+json:
              schema:
                $ref: '#/components/schemas/FeatureCollection'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
...
components:
  parameters:
    SatelliteId:
      name: satellite_id
      in: query
      schema:
        $ref: '#/components/schemas/SatelliteId'
    SatelliteIdPath:
      name: satellite_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/SatelliteId'
    RequestId:
      name: request_id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
    Cursor:
      name: cursor
      in: query
      description: |
        Opaque cursor from a previous response's `next_cursor`. Not a page
        number and not parseable. Offset pagination is not offered: these lists
        grow at the head under active planning, so offsets shift under the
        reader and produce duplicates and gaps.
      schema: { type: string, maxLength: 512 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: |
        The `ETag` of a previously fetched rendering. If it still matches, the
        server answers `304` with no body.

        This is the payload budget's main lever, and it exists because of how
        the globe is actually used: a client polls the same horizon every few
        seconds and the answer changes only when a plan is committed. Re-sending
        an unchanged CZML document each time is the difference between a live
        view and a view that is merely expensive.
      required: false
      schema: { type: string, maxLength: 256 }
  responses:
    BadRequest:
      description: Malformed syntax — unparseable parameter, inverted window, bad cursor.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No such plan, request, or bucket.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    ServiceUnavailable:
      description: |
        The read model cannot be reached. Distinct from an empty result: "there
        is no plan" and "I cannot tell you whether there is a plan" are
        different answers and a client must not conflate them.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotModified:
      description: |
        The rendering is byte-identical to the one the client already has.

        No body, deliberately: a 304 carrying a payload defeats the entire
        point. The `ETag` is repeated so a client that discarded it can recover
        without a second round trip.
      headers:
        ETag:
          $ref: '#/components/headers/ETag'
  headers:
    ETag:
      description: |
        A strong validator over the rendered bytes.

        Strong, not weak: these documents are generated deterministically from
        the read model, so two responses with the same tag really are identical
        octet for octet. A weak tag would allow "semantically equivalent" and
        there is no case here where that buys anything.

        The tag covers the RENDERING, not the read model as a whole. Two
        different query windows over the same data are different documents and
        must not share a tag.
      required: true
      schema: { type: string, maxLength: 256 }
  schemas:
    Uuid:
      type: string
      format: uuid
    SatelliteId:
      type: string
      pattern: '^[A-Z0-9][A-Z0-9_-]{0,31}$'
      minLength: 1
      maxLength: 32
...
    AcquisitionStatus:
      type: string
      enum: [ACTIVE, SUPERSEDED, EXECUTED]
      description: |
        Mirrors the status column the non-overlap constraint is predicated on.
        SUPERSEDED rows are retained history (ADR-0012) and are excluded from
        the live set.
    TimeWindow:
      type: object
      required: [start, end]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
    Staleness:
      type: object
      required: [as_of, lag_seconds]
      description: |
        How current this answer is. Present on every projection response,
        because a read model that cannot say how far behind it is forces the UI
        to guess and the UI guesses optimistically.
      properties:
        as_of:
          type: string
          format: date-time
          description: The occurred_at of the last event folded into this projection.
        lag_seconds:
          type: number
          minimum: 0
          description: Seconds between `as_of` and now, at the moment of the read.
...
    Plan:
      type: object
      required:
        [plan_id, satellite_id, bucket, plan_version, policy, committed_at, acquisitions, staleness]
      properties:
        plan_id: { $ref: '#/components/schemas/Uuid' }
        satellite_id: { $ref: '#/components/schemas/SatelliteId' }
        bucket: { $ref: '#/components/schemas/TimeWindow' }
        plan_version: { type: integer, minimum: 1 }
        supersedes_plan_id: { $ref: '#/components/schemas/Uuid' }
        superseded_by_plan_id:
          $ref: '#/components/schemas/Uuid'
        policy:
          type: string
          description: Which allocation policy produced this plan. Named so a benchmark result can be tied to a plan a reviewer can look at.
        committed_at: { type: string, format: date-time }
        acquisitions:
          type: array
          items:
            $ref: '#/components/schemas/Acquisition'
        metrics:
          type: object
          additionalProperties: true
          description: Plan-level metrics as committed on planning.plan.committed.v1.
        staleness: { $ref: '#/components/schemas/Staleness' }
    PlanPage:
      type: object
      required: [items, staleness]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Plan'
        next_cursor:
          type: string
          description: Absent on the last page.
        staleness: { $ref: '#/components/schemas/Staleness' }
    AcquisitionPage:
      type: object
      required: [items, staleness]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Acquisition'
        next_cursor: { type: string }
        staleness: { $ref: '#/components/schemas/Staleness' }
...
    OpportunityPage:
      type: object
      required: [items, staleness]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Opportunity'
        next_cursor: { type: string }
        staleness: { $ref: '#/components/schemas/Staleness' }
    RequestView:
      type: object
      required: [request_id, state, staleness]
      properties:
        request_id: { $ref: '#/components/schemas/Uuid' }
        customer_id: { type: string }
        target_name: { type: string }
        state:
          type: string
          enum:
            [RECEIVED, AWAITING_PLANNING, PLANNED, ACQUIRED, INFEASIBLE, REJECTED, EXPIRED, CANCELLED]
        window: { $ref: '#/components/schemas/TimeWindow' }
        opportunity_count: { type: integer, minimum: 0 }
        acquisition: { $ref: '#/components/schemas/Acquisition' }
        unfulfilment: { $ref: '#/components/schemas/Unfulfilment' }
        staleness: { $ref: '#/components/schemas/Staleness' }
...
    Polygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              $ref: '#/components/schemas/Position'
    Feature:
      type: object
      required: [type, geometry, properties]
      properties:
        type:
          type: string
          enum: [Feature]
        geometry: { $ref: '#/components/schemas/Polygon' }
        properties:
          type: object
          additionalProperties: true
    FeatureCollection:
      type: object
      required: [type, features, truncated]
      description: |
        RFC 7946 FeatureCollection. Geometry is not re-specified here.

        `truncated` and `staleness` are foreign members, which RFC 7946 §6.1
        permits and every GeoJSON reader ignores. Putting them on the collection
        rather than in a wrapper object keeps the response loadable by deck.gl
        and Cesium directly, with no unwrapping step in the client.
      properties:
        type:
          type: string
          enum: [FeatureCollection]
        features:
          type: array
          items:
            $ref: '#/components/schemas/Feature'
        truncated:
          type: boolean
          description: |
            Whether `limit` cut the result short.

            Explicit, because the alternative is a coverage view that shows a
            gap where there is really a limit — a viewport confidently drawing
            emptiness over a fully-tasked region. A client that sees this set
            must narrow its window rather than believe what it was sent.
        staleness:
          $ref: '#/components/schemas/Staleness'
...
    Problem:
      type: object
      required: [type, title, status]
      description: |
        RFC 9457 Problem Details, the same shape tasking-api uses. Inventing a
        second error envelope for the second service is how APIs become
        inconsistent.
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        reason_code: { type: string }
        correlation_id: { type: string, format: uuid }